Update server and SelectChannel component on channel change
This makes sure that we are using the channels stored in Redux if availible, if not we fall back to what are availible from Django template. Each time a channel is created, we request a new list of channels from the server and update the persisting Redux store.

In turn, we can then update the SelectChannel with this new list, and activate the new option within the list.