Possible fix for channel list becoming out of sync with database
It was possible to get the channels list to become out of sync with the database when relying on Redux to update the list instead of data from Django view. Normally seen when deleting channels from the database after have created some in the UI.

This fix should make sure we always rely on the channels from the database as the primary source of truth. If we need to use Redux to add channels (like when creating them but not moving to a new page) then we should still try and do that.