Additional fix for bug that meant DB and Redux could become out of sync
As a template is refreshed (i.e. coming from a different page), we need to use the channels provided by the database. If we are on the same page but relying on a dynamic change (i.e creating a new channel that in turn updates the channel list component) we need to rely on Redux as it has the most up to date data.

We hit the problem of not knowing if the Redux store has been updated on page reload with Redux persist or if it has been updated with a standard action like adding a new channel.

To get around this, we need to set a stale state on the redux Channels store that will indicate to other components that they need to get the latest from this source instead of the database. In all other cases, use the database.